+++ pyraxshell/globals.py
@@ -11,3 +11,11 @@ LOG_CONF_FILE = '~/.pyraxshell/logging.conf'
 
 # log file
 LOG_FILE = '~/.pyraxshell/pyraxshell.log'
+
+# log levels
+DEBUG = 0
+INFO = INFORMATION = 1
+WARN = WARNING = 2
+ERROR = 3
+CRITICAL = 4
+log_levels = (DEBUG, INFO, WARN, ERROR, CRITICAL)